feat(suspense): mark as experimental and small refactor
