+++ packages/docs/src/routes/api/qwik-optimizer/index.mdx
@@ -116,6 +116,19 @@ Enable the usePreventNavigate hook
 </td></tr>
 <tr><td>
 
+suspense
+
+</td><td>
+
+`"suspense"`
+
+</td><td>
+
+Enable the Suspense fallback primitive
+
+</td></tr>
+<tr><td>
+
 valibot
 
 </td><td>
@@ -1448,6 +1461,8 @@ export type ServerQwikManifest = Pick<
 **References:** [QwikManifest](#qwikmanifest)
 
 [Edit this section](https://github.com/QwikDev/qwik/tree/main/packages/qwik-vite/src/types.ts)
+
+<h2 id="experimentalfeatures-suspense">suspense</h2>
 
 <h2 id="symbolmapper">SymbolMapper</h2>
