+++ packages/qwik/src/core/shared/vnode/vnode-dirty.ts
@@ -1,61 +1,25 @@
 import { isServer } from '@qwik.dev/core/build';
-import { vnode_getProp, vnode_setProp, type VNodeJournal } from '../../client/vnode-utils';
+import { type VNodeJournal } from '../../client/vnode-utils';
 import type { ISsrNode, SSRContainer } from '../../ssr/ssr-types';
 import { addCursor, findCursor, isCursor } from '../cursor/cursor';
 import { getCursorData, type CursorData } from '../cursor/cursor-props';
 import { _executeSsrChores } from '../cursor/ssr-chore-execution';
 import { isServerPlatform } from '../platform/platform';
 import type { Container } from '../types';
-import { QCursorBoundary, QNearestCursorBoundary } from '../utils/markers';
 import { throwErrorAndStop } from '../utils/log';
 import { isPromise } from '../utils/promises';
-import type { CursorBoundary } from '../../use/use-cursor-boundary';
+import {
+  getNearestCursorBoundary,
+  getOwnCursorBoundary,
+  setNearestCursorBoundary,
+  type CursorBoundary,
+} from '../../use/use-cursor-boundary';
 import { ChoreBits } from './enums/chore-bits.enum';
 import type { VNodeOperation } from './types/dom-vnode-operation';
 import type { VNode } from './vnode';
 
 /** Reusable path array to avoid allocations */
 const reusablePath: VNode[] = [];
-
-function getOwnCursorBoundary(container: Container, vNode: VNode): CursorBoundary | null {
-  return container.getHostProp<CursorBoundary>(vNode as any, QCursorBoundary);
-}
-
-function getNearestCursorBoundaryProp(vNode: VNode): CursorBoundary | null {
-  return (
-    (vnode_getProp(vNode, QNearestCursorBoundary, null) as CursorBoundary | null | undefined) ||
-    null
-  );
-}
-
-export function getNearestCursorBoundary(
-  container: Container,
-  vNode: VNode
-): CursorBoundary | null {
-  return getNearestCursorBoundaryProp(vNode) || getOwnCursorBoundary(container, vNode);
-}
-
-function setNearestCursorBoundary(vNode: VNode, boundary: CursorBoundary | null): void {
-  vnode_setProp(vNode, QNearestCursorBoundary, boundary);
-}
-
-export function setVNodeCursorBoundary(
-  container: Container,
-  vNode: VNode,
-  boundary: CursorBoundary | null
-): void {
-  vnode_setProp(vNode, QNearestCursorBoundary, boundary);
-
-  const dirtyChildren = vNode.dirtyChildren;
-  if (!dirtyChildren || dirtyChildren.length === 0) {
-    return;
-  }
-
-  for (let i = 0; i < dirtyChildren.length; i++) {
-    const child = dirtyChildren[i];
-    setVNodeCursorBoundary(container, child, getOwnCursorBoundary(container, child) || boundary);
-  }
-}
 
 /** Propagates CHILDREN dirty bits through the collected path up to the target ancestor */
 function propagatePath(target: VNode): void {
@@ -82,9 +46,11 @@ function propagateToCursorRoot(container: Container, vNode: VNode, cursorRoot: V
   while (current) {
     const isDirty = current.dirty & ChoreBits.DIRTY_MASK;
     const currentIsCursor = isCursor(current);
-    cursorBoundary ||=
-      getOwnCursorBoundary(container, current) ||
-      (isDirty ? getNearestCursorBoundary(container, current) : null);
+    if (__EXPERIMENTAL__.suspense) {
+      cursorBoundary ||=
+        getOwnCursorBoundary(container, current) ||
+        (isDirty ? getNearestCursorBoundary(container, current) : null);
+    }
 
     // Stop when we reach the cursor root or a dirty ancestor
     if (current === cursorRoot || isDirty) {
@@ -122,14 +88,18 @@ function propagateToCursorRoot(container: Container, vNode: VNode, cursorRoot: V
  */
 function findAndPropagateToBlockingCursor(container: Container, vNode: VNode): boolean {
   reusablePath.push(vNode);
-  let cursorBoundary = getOwnCursorBoundary(container, vNode);
+  let cursorBoundary: CursorBoundary | null = __EXPERIMENTAL__.suspense
+    ? getOwnCursorBoundary(container, vNode)
+    : null;
   let current: VNode | null = vNode.slotParent || vNode.parent;
 
   while (current) {
     const currentIsCursor = isCursor(current);
-    cursorBoundary ||=
-      getOwnCursorBoundary(container, current) ||
-      (currentIsCursor ? getNearestCursorBoundary(container, current) : null);
+    if (__EXPERIMENTAL__.suspense) {
+      cursorBoundary ||=
+        getOwnCursorBoundary(container, current) ||
+        (currentIsCursor ? getNearestCursorBoundary(container, current) : null);
+    }
 
     if (currentIsCursor) {
       setNearestCursorBoundary(vNode, cursorBoundary);
